requirements.txt
pymongo==4.13.2
python-dotenv==1.1.1
PyQt6==6.9.1